Started using the planet file API!
I added a unittest for the new planet API - an online file explorer - and I made a main file to test it.
The main file is structured to look like a command line and is entirely for testing purposes.
Through the cmd I was able to see raw polygon data for all places!
Next I'll try to add a display for all this data.