Increased size of screen and added a bounding box to the output of the city finder
Had to fetch another tile as well to make it fit the new larger screen, but it should make for a better experience.
This means now the game is confined to the chosen city instead of allowing you to roam around wherever you want!
Next I'll remove all old code.