Added basic 3d stuff.
I copied a previous project for the basic 3D code using openGL into a `main3d.py`.
This means that users can still use the `main.py` until I've finished with this.
Currently it's only a single cube, but the structure's there.
Next I'll work on movement in the 3D world.