Finally made the API for housing and such work and implemented it in main.py!
This means now I can get on with implementing the new housing prices in the 3d main.
Although it isn't the best numbers (as what I wanted to find was very specific and almost nothing was avaliable without being an index) it works and is a rough estimate and that is all that matters.
Next I'll fix the having 2 mains problem