Moved the 2d example into `main2d.py` and the 3d into `main.py`.
This means that now it is less confusing to users on which to use - I specified in the readme what the files are.
Next I'll work on implementing the new api in main3d.py
It's still slightly confusing due to there being 2 files but I explained it all in the readme so everything is fine.